+++ tests/Functional/Entity/EntityWithContextNested.php
@@ -1,0 +1,23 @@
+<?php
+
+/*
+ * This file is part of the NelmioApiDocBundle package.
+ *
+ * (c) Nelmio
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Nelmio\ApiDocBundle\Tests\Functional\Entity;
+
+use Symfony\Component\Serializer\Attribute\Groups;
+
+class EntityWithContextNested
+{
+    #[Groups(['nested'])]
+    public string $inNestedGroup;
+
+    #[Groups(['other'])]
+    public string $inOtherGroup;
+}