+++ CHANGELOG.md
@@ -1,5 +1,59 @@
 # CHANGELOG
 
+## [5.12.1](https://github.com/nelmio/NelmioApiDocBundle/compare/v5.12.0...v5.12.1) (2026-09-07)
+
+
+### Bug Fixes
+
+* **model-registry:** reserve generated schema names against nested registrations ([#2800](https://github.com/nelmio/NelmioApiDocBundle/issues/2800)) ([3275a03](https://github.com/nelmio/NelmioApiDocBundle/commit/3275a03be706833aa61b7dbd7264a327f9261d40))
+
+
+### Miscellaneous Chores
+
+* **deps:** bump @scalar/api-reference from 1.66.1 to 1.67.0 in /utils ([#2795](https://github.com/nelmio/NelmioApiDocBundle/issues/2795)) ([32bdb0a](https://github.com/nelmio/NelmioApiDocBundle/commit/32bdb0aa56db93d4c5307b048f54ff082ab7fb71))
+* **deps:** bump @stoplight/elements from 9.0.24 to 9.0.25 in /utils ([9066109](https://github.com/nelmio/NelmioApiDocBundle/commit/9066109dc3a2f712668be42cacaa4dfdc73db75d))
+
+## [5.12.0](https://github.com/nelmio/NelmioApiDocBundle/compare/v5.11.1...v5.12.0) (2026-09-03)
+
+
+### Features
+
+* pass #[Context] attribute data to property describers ([#2722](https://github.com/nelmio/NelmioApiDocBundle/issues/2722)) ([f868135](https://github.com/nelmio/NelmioApiDocBundle/commit/f86813575f4e7a0ff98ba2ed487afd1ff41f5091))
+* **type-describer:** add stoppable type describer support ([#2793](https://github.com/nelmio/NelmioApiDocBundle/issues/2793)) ([ee12d2f](https://github.com/nelmio/NelmioApiDocBundle/commit/ee12d2fd16f06108c891d9114cce8bca50ea9e35))
+
+
+### Bug Fixes
+
+* **#2790:** ReflectionReader only substitutes prefixes on methods ([#2791](https://github.com/nelmio/NelmioApiDocBundle/issues/2791)) ([f0f2b01](https://github.com/nelmio/NelmioApiDocBundle/commit/f0f2b01ae665d596811e9c4c571b468f74f030a9))
+* **type-describer:** register autoconfiguration for `TypeDescriberInterface` ([ee12d2f](https://github.com/nelmio/NelmioApiDocBundle/commit/ee12d2fd16f06108c891d9114cce8bca50ea9e35))
+
+## [5.11.1](https://github.com/nelmio/NelmioApiDocBundle/compare/v5.11.0...v5.11.1) (2026-08-12)
+
+
+### Bug Fixes
+
+* harden OpenAPI generation for FrankenPHP / long-lived workers ([#2772](https://github.com/nelmio/NelmioApiDocBundle/issues/2772)) ([7b8ab07](https://github.com/nelmio/NelmioApiDocBundle/commit/7b8ab074d6a01308d96eeebeeb3d682dedadcdd0))
+
+## [5.11.0](https://github.com/nelmio/NelmioApiDocBundle/compare/v5.10.3...v5.11.0) (2026-08-03)
+
+
+### Features
+
+* Add support for the query http method ([#2732](https://github.com/nelmio/NelmioApiDocBundle/issues/2732)) ([a46818f](https://github.com/nelmio/NelmioApiDocBundle/commit/a46818fde0f4351c4628580c909c6e80f9aac852))
+
+
+### Bug Fixes
+
+* **deps:** require zircote/swagger-php &gt;= 5.7.8 for the query method ([#2782](https://github.com/nelmio/NelmioApiDocBundle/issues/2782)) ([1eef066](https://github.com/nelmio/NelmioApiDocBundle/commit/1eef066c3d3b2abd4b491f0dc8ffa357d45b186f))
+
+
+### Miscellaneous Chores
+
+* **deps:** bump @scalar/api-reference from 1.57.5 to 1.62.5 in /utils ([#2769](https://github.com/nelmio/NelmioApiDocBundle/issues/2769)) ([cabb52e](https://github.com/nelmio/NelmioApiDocBundle/commit/cabb52e8406aba7cc93f82a568fa11215ad283ad))
+* **deps:** bump @scalar/api-reference from 1.62.5 to 1.64.0 in /utils ([#2780](https://github.com/nelmio/NelmioApiDocBundle/issues/2780)) ([e2b085c](https://github.com/nelmio/NelmioApiDocBundle/commit/e2b085c7abbe83c10492c6fc8dc990fc44ee4fb0))
+* **deps:** bump @stoplight/elements from 9.0.21 to 9.0.24 in /utils ([#2774](https://github.com/nelmio/NelmioApiDocBundle/issues/2774)) ([c8f3e4e](https://github.com/nelmio/NelmioApiDocBundle/commit/c8f3e4ec890f7ee3ece122926ae936a5df6f72f5))
+* **deps:** bump swagger-ui-dist from 5.32.6 to 5.32.11 in /utils ([#2777](https://github.com/nelmio/NelmioApiDocBundle/issues/2777)) ([9654f06](https://github.com/nelmio/NelmioApiDocBundle/commit/9654f06889e7683a74dffdd21f0048b920894a90))
+
 ## [5.10.3](https://github.com/nelmio/NelmioApiDocBundle/compare/v5.10.2...v5.10.3) (2026-06-01)